MAINT: collect volumes using inotify once again
